+++ core/src/com/boomchess/game/frontend/screen/RelativeResizer.java
@@ -23,7 +23,7 @@ public class RelativeResizer {
          */
         RelativeResizer.width = Gdx.graphics.getWidth();
         RelativeResizer.height = Gdx.graphics.getHeight();
-        BoomChess.tileSize = (int) (width / 20);
+        BoomChess.tileSize = width / 20;
     }
 
     // this function should be run each frame
